Core: rename geometry identifier file
