+++ Examples/Algorithms/MaterialMapping/include/ActsExamples/MaterialMapping/IMaterialWriter.hpp
@@ -8,7 +8,7 @@
 
 #pragma once
 
-#include "Acts/Geometry/GeometryID.hpp"
+#include "Acts/Geometry/GeometryIdentifier.hpp"
 
 #include <map>
