+++ Examples/Algorithms/TrackFinding/src/TrackFindingOptions.cpp
@@ -8,7 +8,7 @@
 
 #include "ActsExamples/TrackFinding/TrackFindingOptions.hpp"
 
-#include "Acts/Geometry/GeometryID.hpp"
+#include "Acts/Geometry/GeometryIdentifier.hpp"
 
 #include <string>
