+++ Examples/Io/Json/include/ActsExamples/Plugins/Json/JsonMaterialWriter.hpp
@@ -12,7 +12,7 @@
 
 #pragma once
 
-#include "Acts/Geometry/GeometryID.hpp"
+#include "Acts/Geometry/GeometryIdentifier.hpp"
 #include "Acts/Material/ISurfaceMaterial.hpp"
 #include "Acts/Material/IVolumeMaterial.hpp"
 #include "Acts/Plugins/Json/JsonGeometryConverter.hpp"
