+++ Examples/Io/Json/src/JsonMaterialWriter.cpp
@@ -8,7 +8,7 @@
 
 #include "ActsExamples/Plugins/Json/JsonMaterialWriter.hpp"
 
-#include "Acts/Geometry/GeometryID.hpp"
+#include "Acts/Geometry/GeometryIdentifier.hpp"
 #include "Acts/Material/BinnedSurfaceMaterial.hpp"
 
 #include <fstream>
