+++ Examples/Io/Root/include/ActsExamples/Io/Root/RootMaterialDecorator.hpp
@@ -9,7 +9,7 @@
 #pragma once
 
 #include "ActsExamples/Framework/ProcessCode.hpp"
-#include <Acts/Geometry/GeometryID.hpp>
+#include <Acts/Geometry/GeometryIdentifier.hpp>
 #include <Acts/Geometry/TrackingVolume.hpp>
 #include <Acts/Material/IMaterialDecorator.hpp>
 #include <Acts/Material/ISurfaceMaterial.hpp>
