+++ Examples/Io/Root/src/RootMaterialTrackWriter.cpp
@@ -8,7 +8,7 @@
 
 #include "ActsExamples/Io/Root/RootMaterialTrackWriter.hpp"
 
-#include <Acts/Geometry/GeometryID.hpp>
+#include <Acts/Geometry/GeometryIdentifier.hpp>
 #include <Acts/Surfaces/CylinderBounds.hpp>
 #include <Acts/Surfaces/RadialBounds.hpp>
 #include <Acts/Utilities/Helpers.hpp>
