+++ Examples/Run/Fatras/Common/FatrasSimulationBase.cpp
@@ -8,7 +8,7 @@
 
 #include "FatrasSimulationBase.hpp"
 
-#include "Acts/Geometry/GeometryID.hpp"
+#include "Acts/Geometry/GeometryIdentifier.hpp"
 #include "Acts/Geometry/TrackingGeometry.hpp"
 #include "Acts/MagneticField/ConstantBField.hpp"
 #include "Acts/MagneticField/InterpolatedBFieldMap.hpp"
