+++ Tests/UnitTests/Core/Geometry/GeometryIDTests.cpp
@@ -8,7 +8,7 @@
 
 #include <boost/test/unit_test.hpp>
 
-#include "Acts/Geometry/GeometryID.hpp"
+#include "Acts/Geometry/GeometryIdentifier.hpp"
 
 namespace Acts {
 namespace Test {
